Background thread writer
