+++ src/ext/coms_curl.h
@@ -1,0 +1,7 @@
+#ifndef DD_COMS_CURL_H
+#define DD_COMS_CURL_H
+
+uint32_t curl_ze_data_out();
+
+
+#endif // DD_COMS_CURL_H
